Created "LoadDatabase" class
